добавил изменения к треугольнику и сделал программу на проверку битых ссылок

+++ lab1/Triangle/Triangle/FuctionsForTriangle.h
@@ -1,6 +1,6 @@
 #pragma once
 bool StringIsFloat(string vertexInString);
-bool TriangleExists(float a, float b, float c);
-bool IsEquilateralTriangle(float a, float b, float c);
-bool IsIsoscelesTriangle(float a, float b, float c);
+bool TriangleExists(double a, double b, double c);
+bool IsEquilateralTriangle(double a, double b, double c);
+bool IsIsoscelesTriangle(double a, double b, double c);
